Extract FeePresenter
