+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swap/approve/SwapApproveViewModel.kt
@@ -3,11 +3,8 @@ package io.horizontalsystems.bankwallet.modules.swap.approve
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import io.horizontalsystems.bankwallet.core.App
-import io.horizontalsystems.bankwallet.entities.Coin
-import io.horizontalsystems.bankwallet.modules.guides.DataState
 import io.horizontalsystems.core.SingleLiveEvent
 import io.reactivex.disposables.CompositeDisposable
-import kotlin.math.min
 
 class SwapApproveViewModel(private val service: ISwapApproveService) : ViewModel() {
 
@@ -19,39 +16,13 @@ class SwapApproveViewModel(private val service: ISwapApproveService) : ViewModel
 
     private val disposables = CompositeDisposable()
 
-    val feeValue = MutableLiveData<String>()
-    val feeLoading = MutableLiveData<Boolean>()
     val approveAllowed = MutableLiveData<Boolean>()
     val closeLiveEvent = SingleLiveEvent<Unit>()
     val errorLiveEvent = SingleLiveEvent<Throwable>()
 
+    val feePresenter = FeePresenter(service.feeService)
+
     init {
-        service.feeValues
-                .subscribe {
-                    feeLoading.postValue(it is DataState.Loading)
-
-                    when (it) {
-                        is DataState.Success -> {
-                            val coinValue = it.data.first
-
-                            var res = App.numberFormatter.formatCoin(coinValue.value, coinValue.coin.code, 0, min(coinValue.coin.decimal, 8))
-
-                            it.data.second?.let { fiatValue ->
-                                res += " | "
-                                res += App.numberFormatter.formatFiat(fiatValue.value, fiatValue.currency.symbol, 0, 2)
-                            }
-
-                            feeValue.postValue(res)
-                        }
-                        is DataState.Error -> {
-                            errorLiveEvent.postValue(it.throwable)
-                        }
-                    }
-                }
-                .let {
-                    disposables.add(it)
-                }
-
         service.approveState
                 .subscribe {
                     approveAllowed.postValue(it is SwapApproveState.ApproveAllowed)
